Extract search bar actions into a stabilized connected component
